requirements.txt
pymupdf
sentence-transformers
faiss-cpu
requests
streamlit
